Added ability to specify reason for editing or deleting events

+++ calendarapp/event_views.py
@@ -11,7 +11,7 @@ import time
 from django.template.loader import render_to_string
 from .tokens import account_activation_token
 from .models import Year, Month, Day, Calendar, Event, Location, DayOfWeek
-from .forms import EventForm, CalendarForm, MemberCreationForm, MemberChangeForm
+from .forms import EventForm, CalendarForm, MemberCreationForm, MemberChangeForm, ReasonForEventEditForm, ReasonForEventRejectForm, ReasonForEventDeleteForm, ReasonForEventDeleteRejectForm
 from django.conf import settings
 
 def event_conflicts(test_event, original_event=None):
@@ -32,6 +32,10 @@ def event_conflicts(test_event, original_event=None):
 	if len(conflicts) == 0:
 		return None
 	return conflicts
+
+def calendarEventView(request, year, month, day, pk, slug):
+	event = get_object_or_404(Event, pk=pk, slug=slug)
+	return render(request, 'event_view.html', {'event': event})
 
 def newEventView(request):
 	if request.method == 'POST':
@@ -65,7 +69,6 @@ def newEventView(request):
 				msg.attach_alternative(html_message, 'text/html')
 				msg.send()
 				return redirect('event_approval_sent', slug=new_event.slug, pk=new_event.pk)
-
 	else:
 		start_time = timezone.localtime()
 		start_time = start_time - datetime.timedelta(seconds=start_time.minute*60)
@@ -90,7 +93,7 @@ def approve_event(request, slug, pk):
 	event = get_object_or_404(Event, slug=slug, pk=pk)
 	if Group.objects.get(name='Admins') in request.user.groups.all():
 		#Send an email to the user if their email is confirmed to tell them about the event
-		if event.creator.member.email_confirmed and event.approved == False:
+		if event.approved == False:
 			event.approved = True
 			event.save()
 			current_site = get_current_site(request)
@@ -112,37 +115,43 @@ def approve_event(request, slug, pk):
 		return render(request, 'approve/event_approval_error.html')
 
 def reject_event(request, slug, pk):
-	try:
-		event = Event.objects.get(slug=slug, pk=pk)
-	except Event.DoesNotExist:
-		return render(request, 'approve/event_already_rejected.html')
 	if Group.objects.get(name='Admins') in request.user.groups.all():
-		#Send an email to the user if their email is confirmed to tell them about the event
-		if event.creator.member.email_confirmed:
-			current_site = get_current_site(request)
-			subject = f'Event Rejected: {event.title} on {event.start_date}'
-			content = {
-				'user': event.creator,
-				'event': event,
-			}
-			text_message = render_to_string('email/event_rejected_email.html', content)
-			html_message = render_to_string('email/event_rejected_html_email.html', content)
-			msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, [event.creator.email])
-			msg.attach_alternative(html_message, 'text/html')
-			msg.send()
-		event.delete()
-		return redirect('month', year=event.start_date.year, month=event.start_date.month)
-	else:
-		return render(request, 'approve/event_approval_error.html')
-
-def calendarEventView(request, year, month, day, pk, slug):
-	event = get_object_or_404(Event, pk=pk, slug=slug)
-	return render(request, 'event_view.html', {'event': event})
+		if request.method == 'POST':
+			reason_form = ReasonForEventRejectForm(request.POST)
+			if reason_form.is_valid():
+				try:
+					event = Event.objects.get(slug=slug, pk=pk)
+				except Event.DoesNotExist:
+					return render(request, 'approve/event_already_rejected.html')
+				#Send an email to the user if their email is confirmed to tell them about the event
+				reason = reason_form.cleaned_data.get('reason')
+				current_site = get_current_site(request)
+				subject = f'Event Rejected: {event.title} on {event.start_date}'
+				content = {
+					'user': event.creator,
+					'event': event,
+					'reason': reason,
+				}
+				text_message = render_to_string('email/event_rejected_email.html', content)
+				html_message = render_to_string('email/event_rejected_html_email.html', content)
+				msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, [event.creator.email])
+				msg.attach_alternative(html_message, 'text/html')
+				msg.send()
+				event.delete()
+				return redirect('month', year=event.start_date.year, month=event.start_date.month)
+		else:
+			reason_form = ReasonForEventRejectForm()
+			return render(request, 'approve/reason_for_reject.html', {'reason_form': reason_form})
+	return render(request, 'approve/event_approval_error.html')
 
 def editEventView(request, year, month, day, pk, slug):
 	event = get_object_or_404(Event, pk=pk, slug=slug)
 	if request.method == 'POST':
 		event_form = EventForm(request.POST, instance=event)
+		reason_form = ReasonForEventEditForm(request.POST)
+		if reason_form:
+			if reason_form.is_valid():
+				reason = reason_form.cleaned_data.get('reason')
 		if event_form.is_valid():
 			if event_form.has_changed():
 				changed_event = event_form.save(commit=False)
@@ -164,14 +173,15 @@ def editEventView(request, year, month, day, pk, slug):
 				else:
 					#Send email to admins if event edited by non-admin
 					current_site = get_current_site(request)
-					changed_event_conflicts = event_conflicts(changed_event, original_event)
 					subject = f'Approve Event Edit: {changed_event.title} on {changed_event.start_date}'
 					content = {
 						'user': request.user,
 						'domain': current_site.domain,
 						'changed_event': changed_event,
-						'changed_event_conflicts': changed_event_conflicts,
+						'changed_event_conflicts': event_conflicts(changed_event, original_event),
 						'original_event': original_event,
+						'original_event_conflicts': event_conflicts(original_event),
+						'reason': reason,
 					}
 					text_message = render_to_string('email/approve_event_edit_email.html', content)
 					html_message = render_to_string('email/approve_event_edit_html_email.html', content)
@@ -191,13 +201,12 @@ def editEventView(request, year, month, day, pk, slug):
 				return redirect('month', year=event.start_date.year, month=event.start_date.month)
 	else:
 		event_form = EventForm(instance=event)
+		if Group.objects.get(name='Admins') in request.user.groups.all():
+			reason_form = None
+		else:
+			reason_form = ReasonForEventEditForm()
 		#Creates copy of event and stores it in session data
 		if 'original_event_pk' in request.session:
-			#Delete the event stored in the original_event key
-			try:
-				Event.objects.get(pk=request.session['original_event_pk']).delete()
-			except Event.DoesNotExist:
-				pass
 			#Delete the key
 			del request.session['original_event_pk']
 			#Create event copy
@@ -209,7 +218,6 @@ def editEventView(request, year, month, day, pk, slug):
 			event.save()
 			#Store new copy event in the session data
 			request.session['original_event_pk'] = event.pk
-			print(request.session['original_event_pk'])
 		else:
 			#If there is no key, create copy event and store it in sessions
 			event.pk = None
@@ -218,24 +226,16 @@ def editEventView(request, year, month, day, pk, slug):
 			event.approved = False
 			event.save()
 			request.session['original_event_pk'] = event.pk
-			print(request.session['original_event_pk'])
 			
-	return render(request, 'edit_event.html', {'event_form': event_form})
-
-def deleteEventView(request, year, month, day, pk, slug):
-	event = get_object_or_404(Event, pk=pk, slug=slug)
-	if request.method == 'POST':
-		event.delete()
-		return redirect('month', year=timezone.now().year, month=timezone.now().month)
-	return render(request, 'delete_event.html', {'event': event})
+	return render(request, 'edit_event.html', {'event_form': event_form, 'reason_form': reason_form})
 
 def approve_event_change(request, original_slug, original_pk, changed_slug, changed_pk):
-	try:
-		Event.objects.get(slug=original_slug, pk=original_pk).delete()
-		changed_event = Event.objects.get(slug=changed_slug, pk=changed_pk)
-	except Event.DoesNotExist:
-		return render(request, 'approve/event_already_approved.html')
 	if Group.objects.get(name='Admins') in request.user.groups.all():
+		try:
+			Event.objects.get(slug=original_slug, pk=original_pk).delete()
+		except Event.DoesNotExist:
+			return render(request, 'approve/event_already_approved.html')
+		changed_event = get_object_or_404(Event, slug=changed_slug, pk=changed_pk)
 		changed_event.approved = True
 		changed_event.save()
 		#Send an email to the user if their event change is confirmed to tell them about the event
@@ -255,27 +255,134 @@ def approve_event_change(request, original_slug, original_pk, changed_slug, chan
 		return render(request, 'approve/event_approval_error.html')
 
 def reject_event_change(request, original_slug, original_pk, changed_slug, changed_pk):
-	try:
-		changed_event = Event.objects.get(slug=changed_slug, pk=changed_pk)
-		original_event = Event.objects.get(slug=original_slug, pk=original_pk)
-	except Event.DoesNotExist:
-		return render(request, 'approve/event_already_rejected.html')
 	if Group.objects.get(name='Admins') in request.user.groups.all():
-		original_event.approved = True
-		original_event.save()
-		#Send an email to the user if their event is denied
-		subject = f'Event Change Rejected: {changed_event.title} on {changed_event.start_date}'
-		content = {'event': changed_event}
-		text_message = render_to_string('email/event_change_rejected_email.html', content)
-		html_message = render_to_string('email/event_change_rejected_html_email.html', content)
-		recipients = []
-		if changed_event.editor != changed_event.creator:
-			recipients.append(changed_event.creator.email)
-		recipients.append(changed_event.editor.email)
-		msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, recipients)
-		msg.attach_alternative(html_message, 'text/html')
-		msg.send()
-		changed_event.delete()
-		return redirect('month', year=event.start_date.year, month=event.start_date.month)
+		if request.method == 'POST':
+			reason_form = ReasonForEventRejectForm(request.POST)
+			if reason_form.is_valid():
+				try:
+					changed_event = Event.objects.get(slug=changed_slug, pk=changed_pk)
+					original_event = Event.objects.get(slug=original_slug, pk=original_pk)
+				except Event.DoesNotExist:
+					return render(request, 'approve/event_already_rejected.html')
+				original_event.approved = True
+				original_event.save()
+				#Send an email to the user if their event is denied
+				reason = reason_form.cleaned_data.get('reason')
+				subject = f'Event Change Rejected: {changed_event.title} on {changed_event.start_date}'
+				content = {
+					'event': changed_event,
+					'reason': reason,
+				}
+				text_message = render_to_string('email/event_change_rejected_email.html', content)
+				html_message = render_to_string('email/event_change_rejected_html_email.html', content)
+				msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, [changed_event.editor.email])
+				msg.attach_alternative(html_message, 'text/html')
+				msg.send()
+				changed_event.delete()
+				return redirect('month', year=original_event.start_date.year, month=original_event.start_date.month)
+		else:
+			reason_form = ReasonForEventRejectForm()
+			return render(request, 'approve/reason_for_reject.html', {'reason_form': reason_form})
+	return render(request, 'approve/event_approval_error.html')
+
+def deleteEventView(request, year, month, day, pk, slug):
+	event = get_object_or_404(Event, pk=pk, slug=slug)
+	if request.method == 'POST':
+		if Group.objects.get(name='Admins') in request.user.groups.all():
+			event.delete()
+			return redirect('month', year=timezone.now().year, month=timezone.now().month)
+		else:
+			reason_form = ReasonForEventDeleteForm(request.POST)
+			if reason_form.is_valid():
+				#Send email to admins if event edited by non-admin
+				reason = reason_form.cleaned_data.get('reason')
+				current_site = get_current_site(request)
+				subject = f'Delete Event: {event.title} on {event.start_date}'
+				content = {
+					'user': request.user,
+					'domain': current_site.domain,
+					'event': event,
+					'event_conflicts': event_conflicts(event),
+					'reason': reason,
+				}
+				text_message = render_to_string('email/approve_event_delete_email.html', content)
+				html_message = render_to_string('email/approve_event_delete_html_email.html', content)
+				recipient_list = []
+				for admin in User.objects.filter(groups__name='Admins'):
+					recipient_list.append(admin.email)
+				msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, recipient_list)
+				msg.attach_alternative(html_message, 'text/html')
+				msg.send()
+				return redirect('event_deletion_request_sent', slug=event.slug, pk=event.pk)
+	else:
+		if Group.objects.get(name='Admins') in request.user.groups.all():
+			reason_form = None
+		else:
+			reason_form = ReasonForEventDeleteForm()
+	return render(request, 'delete_event.html', {'event': event, 'reason_form': reason_form})
+
+def event_deletion_request_sent(request, slug, pk):
+	return render(request, 'approve/event_deletion_request_sent.html')
+
+def approve_event_delete(request, slug, pk, deleting_user_pk):
+	if Group.objects.get(name='Admins') in request.user.groups.all():
+		if request.method == 'POST':
+			reason_form = ReasonForEventDeleteForm(request.POST)
+			if reason_form.is_valid():
+				try:
+					event = Event.objects.get(slug=slug, pk=pk)
+				except Event.DoesNotExist:
+					return render(request, 'approve/event_already_deleted.html')
+				reason = reason_form.cleaned_data.get('reason')
+				deleting_user = get_object_or_404(User, pk=deleting_user_pk)
+				subject = f'Event Deleted: {event.title} on {event.start_date}'
+				content = {
+					'deleting_user': deleting_user,
+					'event': event,
+					'reason': reason,
+				}
+				text_message = render_to_string('email/event_delete_approved_email.html', content)
+				html_message = render_to_string('email/event_delete_approved_html_email.html', content)
+				recipients = []
+				if event.creator != deleting_user:
+					recipients.append(event.creator.email)
+				recipients.append(deleting_user.email)
+				msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, recipients)
+				msg.attach_alternative(html_message, 'text/html')
+				msg.send()
+				event.delete()
+				return redirect('month', year=event.start_date.year, month=event.start_date.month)
+		else:
+			reason_form = ReasonForEventDeleteForm()
+			return render(request, 'approve/reason_for_delete.html', {'reason_form': reason_form})
+	else:
+		return render(request, 'approve/event_approval_error.html')
+
+def reject_event_delete(request, slug, pk, deleting_user_pk):
+	if Group.objects.get(name='Admins') in request.user.groups.all():
+		if request.method == 'POST':
+			reason_form = ReasonForEventDeleteRejectForm(request.POST)
+			if reason_form.is_valid():
+				try:
+					event = Event.objects.get(slug=slug, pk=pk)
+				except Event.DoesNotExist:
+					return render(request, 'approve/event_already_deleted.html')
+				reason = reason_form.cleaned_data.get('reason')
+				deleting_user = get_object_or_404(User, pk=deleting_user_pk)
+				subject = f'Event Not Deleted: {event.title} on {event.start_date}'
+				content = {
+					'deleting_user': deleting_user,
+					'event': event,
+					'reason': reason,
+				}
+				text_message = render_to_string('email/event_delete_rejected_email.html', content)
+				html_message = render_to_string('email/event_delete_rejected_html_email.html', content)
+				msg = EmailMultiAlternatives(subject, text_message, settings.DEFAULT_FROM_EMAIL, [deleting_user.email])
+				msg.attach_alternative(html_message, 'text/html')
+				msg.send()
+				return redirect('month', year=event.start_date.year, month=event.start_date.month)
+		else:
+			reason_form = ReasonForEventDeleteRejectForm()
+			return render(request, 'approve/reason_for_reject.html', {'reason_form': reason_form})
 	else:
 		return render(request, 'approve/event_approval_error.html')
